requirements.txt
aiohttp==3.9.0
aiosignal==1.3.1
anyio==4.0.0
asn1crypto==1.5.1
async-timeout==4.0.3
attrs==23.1.0
base58==2.1.1
bitarray==2.8.3
certifi==2023.11.17
cffi==1.16.0
charset-normalizer==3.3.2
coincurve==18.0.0
coverage==7.3.2
cytoolz==0.12.2
eth-abi==4.2.1
eth-account==0.10.0
eth-hash==0.5.2
eth-keyfile==0.6.1
eth-keys==0.4.0
eth-rlp==0.3.0
eth-typing==3.5.2
eth-utils==2.3.1
exceptiongroup==1.1.3
fastecdsa==2.3.0
frozenlist==1.4.0
greenlet==3.0.1
h11==0.14.0
hexbytes==0.3.1
httpcore==1.0.2
httpx==0.25.1
idna==3.4
iniconfig==2.0.0
jsonschema==4.20.0
jsonschema-specifications==2023.11.1
lru-dict==1.2.0
multidict==6.0.4
numpy==1.26.2
packaging==23.2
parsimonious==0.9.0
pluggy==1.3.0
protobuf==4.25.1
pycparser==2.21
pycryptodome==3.19.0
pytest==7.4.3
pytest-cov==4.1.0
python-dotenv==1.0.0
pyunormalize==15.1.0
referencing==0.31.0
regex==2023.10.3
requests==2.31.0
rlp==3.0.0
rpds-py==0.13.0
sniffio==1.3.0
SQLAlchemy==2.0.23
tomli==2.0.1
toolz==0.12.0
typing_extensions==4.8.0
urllib3==2.1.0
web3==6.11.3
websockets==12.0
yarl==1.9.2
